# gin-vue-admin/server/docs/swagger.yaml
definitions:
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  starNote.StarTag:
    properties:
      ID:
        description: 主键ID
        type: integer
      color:
        description: 颜色
        type: string
      createdAt:
        description: 创建时间
        type: string
      name:
        description: 标签名
        type: string
      updatedAt:
        description: 更新时间
        type: string
    required:
    - color
    - name
    type: object
...
  system.SysError:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      form:
        description: 错误来源
        type: string
      info:
        description: 错误内容
        type: string
      level:
        type: string
      solution:
        description: 解决方案
        type: string
      status:
        description: 处理状态：未处理/处理中/处理完成
        type: string
      updatedAt:
        description: 更新时间
        type: string
    required:
    - form
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.9.0
paths:
...
  /st/createStarTag:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建用户标签
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/starNote.StarTag'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建用户标签
      tags:
      - StarTag
  /st/deleteStarTag:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除用户标签
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/starNote.StarTag'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除用户标签
      tags:
      - StarTag
  /st/deleteStarTagByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除用户标签
      tags:
      - StarTag
  /st/findStarTag:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询用户标签
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/starNote.StarTag'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询用户标签
      tags:
      - StarTag
  /st/getStarTagList:
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: createdAtRange
        type: array
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: order
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取用户标签列表
      tags:
      - StarTag
...
  /st/updateStarTag:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新用户标签
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/starNote.StarTag'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户标签
      tags:
      - StarTag
...
  /sysError/createSysError:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建错误日志
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysError'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建错误日志
      tags:
      - SysError
  /sysError/deleteSysError:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除错误日志
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysError'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除错误日志
      tags:
      - SysError
  /sysError/deleteSysErrorByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除错误日志
      tags:
      - SysError
  /sysError/findSysError:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询错误日志
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.SysError'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询错误日志
      tags:
      - SysError
  /sysError/getSysErrorList:
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: createdAtRange
        type: array
      - in: query
        name: form
        type: string
      - in: query
        name: info
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取错误日志列表
      tags:
      - SysError
  /sysError/updateSysError:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新错误日志
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysError'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新错误日志
      tags:
      - SysError
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...